move mailer functions into own folder
