+++ backend/src/graphql/resolver/UserResolver.ts
@@ -20,7 +20,7 @@ import { UserRepository } from '../../typeorm/repository/User'
 import { LoginUser } from '@entity/LoginUser'
 import { LoginUserBackup } from '@entity/LoginUserBackup'
 import { LoginEmailOptIn } from '@entity/LoginEmailOptIn'
-import { sendAccountActivationEmail, sendEMail } from '../../util/sendEMail'
+import { sendAccountActivationEmail, sendEMail } from '../../mailer/sendEMail'
 import { LoginElopageBuysRepository } from '../../typeorm/repository/LoginElopageBuys'
 import { signIn } from '../../apis/KlicktippController'
 import { RIGHTS } from '../../auth/RIGHTS'
